# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Wood Trading API",
        "description": "API for Wood Trading Application - A comprehensive system for managing wood inventory, sales, purchases, and financial operations.",
        "contact": {
            "name": "Wood Trading Support",
            "email": "support@woodtrading.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local Development Server"
        },
        {
            "url": "https://api.woodtrading.com/api",
            "description": "Production Server"
        },
        {
            "url": "D:\\Haleem-dev\\wood-trading"
        }
    ],
    "paths": {
...
        "/api/customers": {
            "get": {
                "tags": [
                    "Customers"
                ],
                "summary": "Get list of customers",
                "operationId": "1b873d7b0c63876f30a978b52c4d24d2",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search term for name or contact person",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Customers retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/CustomerResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Customers"
                ],
                "summary": "Create a new customer",
                "operationId": "747630c757eae4c38404384330e68c57",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Construction Corp."
                                    },
                                    "contact_person": {
                                        "type": "string",
                                        "example": "John Builder"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john@constructioncorp.com"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "+1-555-2000"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "500 Build Street, Construction City, CC 50000"
                                    },
                                    "tax_id": {
                                        "type": "string",
                                        "example": "CUST001"
                                    },
                                    "credit_limit": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 100000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Customer created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Customer created successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/CustomerResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/customers/{id}": {
            "get": {
                "tags": [
                    "Customers"
                ],
                "summary": "Get customer by ID",
                "operationId": "d03d1278b7e26322e9a425406bc4e9f2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Customer retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/CustomerResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Customers"
                ],
                "summary": "Update customer",
                "operationId": "32da1527afa82d1a6b0e034146f1032d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Updated Construction Corp."
                                    },
                                    "contact_person": {
                                        "type": "string",
                                        "example": "John Builder"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john@constructioncorp.com"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "+1-555-2000"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "500 Build Street, Construction City, CC 50000"
                                    },
                                    "tax_id": {
                                        "type": "string",
                                        "example": "CUST001"
                                    },
                                    "credit_limit": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 100000
                                    },
                                    "is_active": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Customer updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Customer updated successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/CustomerResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Customers"
                ],
                "summary": "Delete customer",
                "operationId": "e9f16ac94d63309225711a1f3dfb245a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Customer deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Customer deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "CustomerResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "ABC Construction Ltd"
                    },
                    "contact_person": {
                        "type": "string",
                        "example": "John Smith"
                    },
                    "email": {
                        "type": "string",
                        "example": "john@abcconstruction.com"
                    },
                    "phone": {
                        "type": "string",
                        "example": "+1-555-0123"
                    },
                    "address": {
                        "type": "string",
                        "example": "123 Main St, City, State"
                    },
                    "tax_id": {
                        "type": "string",
                        "example": "TAX123456789"
                    },
                    "credit_limit": {
                        "type": "number",
                        "format": "float",
                        "example": 50000
                    },
                    "outstanding_balance": {
                        "type": "number",
                        "format": "float",
                        "example": 2500
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "quotations": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/QuotationResource"
                        }
                    },
                    "sales_orders": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/SalesOrderResource"
                        }
                    },
                    "invoices": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/InvoiceResource"
                        }
                    },
                    "payments": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/PaymentResource"
                        }
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "InvoiceResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "invoice_no": {
                        "type": "string",
                        "example": "INV-2024-001"
                    },
                    "customer_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "invoice_date": {
                        "type": "string",
                        "format": "date"
                    },
                    "due_date": {
                        "type": "string",
                        "format": "date"
                    },
                    "subtotal": {
                        "type": "number",
                        "format": "float",
                        "example": 2000
                    },
                    "tax_amount": {
                        "type": "number",
                        "format": "float",
                        "example": 200
                    },
                    "total_amount": {
                        "type": "number",
                        "format": "float",
                        "example": 2200
                    },
                    "paid_amount": {
                        "type": "number",
                        "format": "float",
                        "example": 0
                    },
                    "balance_due": {
                        "type": "number",
                        "format": "float",
                        "example": 2200
                    },
                    "status": {
                        "type": "string",
                        "example": "pending"
                    },
                    "notes": {
                        "type": "string",
                        "example": "Payment due within 30 days"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "PaymentResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "payment_no": {
                        "type": "string",
                        "example": "PAY-2024-001"
                    },
                    "customer_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "invoice_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "payment_date": {
                        "type": "string",
                        "format": "date"
                    },
                    "amount": {
                        "type": "number",
                        "format": "float",
                        "example": 1000
                    },
                    "payment_method": {
                        "type": "string",
                        "example": "bank_transfer"
                    },
                    "reference_no": {
                        "type": "string",
                        "example": "TXN123456789"
                    },
                    "notes": {
                        "type": "string",
                        "example": "Payment for invoice INV-2024-001"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "QuotationResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "quotation_no": {
                        "type": "string",
                        "example": "QUO-2024-001"
                    },
                    "customer_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "quotation_date": {
                        "type": "string",
                        "format": "date"
                    },
                    "valid_until": {
                        "type": "string",
                        "format": "date"
                    },
                    "subtotal": {
                        "type": "number",
                        "format": "float",
                        "example": 1000
                    },
                    "tax_amount": {
                        "type": "number",
                        "format": "float",
                        "example": 100
                    },
                    "total_amount": {
                        "type": "number",
                        "format": "float",
                        "example": 1100
                    },
                    "status": {
                        "type": "string",
                        "example": "pending"
                    },
                    "notes": {
                        "type": "string",
                        "example": "Special pricing for bulk order"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "SalesOrderResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "order_no": {
                        "type": "string",
                        "example": "SO-2024-001"
                    },
                    "customer_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "order_date": {
                        "type": "string",
                        "format": "date"
                    },
                    "delivery_date": {
                        "type": "string",
                        "format": "date"
                    },
                    "subtotal": {
                        "type": "number",
                        "format": "float",
                        "example": 2000
                    },
                    "tax_amount": {
                        "type": "number",
                        "format": "float",
                        "example": 200
                    },
                    "total_amount": {
                        "type": "number",
                        "format": "float",
                        "example": 2200
                    },
                    "status": {
                        "type": "string",
                        "example": "pending"
                    },
                    "notes": {
                        "type": "string",
                        "example": "Rush order - deliver ASAP"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter token in the format: Bearer {token}",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}